# PERF: avoid unnecessary copies factorize