@jbrockmendel 2022-02-22
> Might be good to follow up with a whatsnew note

sure.  will add that to the next "assorted" branch